ft(get-all-question-api): implement get all question api endpoint
- add route
- add tests

[Finishes #159980549]
